Add data vs serialized example fields
This adds two fields to the Example Object and partially
deprecates the existing `value` field.

`dataValue` applies to the data that would be passed to
schema validation.

`serializedValue` (which MUST be a string), like `externalValue`,
applies to the serialized form.

Guidance is provided that `value` (and the shorthand singluar
`example`) are safe for JSON serialization targets and for strings
serialized to targets that do not apply further escaping, but are
otherwise deprecated due to ambiguous behavior.