Add data vs serialized example fields
This adds four fields to the Example Object.

`dataValue` and `externalDataValue` apply to the data that would
be passed to schema validation.

`serializedValue` (which MUST be a string) and `externalSerializedValue`
apply to the serialized form.

External values are used when the appropriate value cannot be
included in JSON or YAML,