Provide parsing and serialization guidance
This creates a "Working with Data" section that
incorporates the existing "Data Types" section (with some
section level adjustments) along with new guidance on mapping
different kinds of data between serialized, data, and application
forms.  This terminology matches the terminology currently
being considered for examples.

The application form is largely out of scope for the OAS, and is
mainly included to clarify this scope while acknowledging that the
OAS may influence such things.

Most of the new material is on parsing and serializing, briefly
addressing JSON as the common case before going into detail
on non-JSON data, with examples.  This is where the requirements
for schema and/or instance inspection/searching are listed.

The only additional change is no longer mentioning the property
schema in the Encoding Object, in part because with the new
`multipart/mixed` support Encoding Objects can be used with
arrays as well as objects.