Fix deepObject explode default behavior.
The `explode` field defaults to `false` with `style: deepObject`,
but the behavior of that combination is undefined in 3.0 and 3.1.

For 3.2, define it to behave the same as
`style: deepObject, explode: true`, which maintains compatibility
(by defining previously undefined behavior rather than changing
the defined default value) and produces intuitive results.