Add style: cookie, explain pct-encoding
This adds a `style: cookie` to provide a non-percent-encoding
option for managing cookie parameters.

It also explains the conditions under which automatic percent-encoding
with `style: form` might work, and also links to a draft update of
the cookie RFC that explains that things that look like percent-encoding
should _not_ be decoded for storage, which means that `style: cookie`
is providing the most compliant behavior for that updated guidance.