Streamline OpenAPI Description Structure post-moves
A lot of what was here is no longer needed, and other things can
be said better either because of the new arrangement or because
I thought of better wording since I first wrote things.

In particular, get rid of the awkward Document vs document distinction
by banishing discussion of small-d documents to the appendix.