Update Media Type Regisry for reorg, review feedback
This updates for various moved and re-titled sections to which
we were linking.  It also replaces the immediate-parent link
with a link to the Object section in which the subsection appears,
since after the section reorg almost everything is under an Object.
This is more robust to minor changes in subsection organization,
but also helps users find the relevant part of the OAS.